# be/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mobile App Backend API",
    "description": "REST API for root info, health check, and user management.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://20.255.57.186:8080",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": ["categories"],
        "summary": "Get root categories",
        "operationId": "getRootCategories",
        "responses": {
          "200": {
            "description": "Root categories retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["categories"],
        "summary": "Create a category",
        "operationId": "createCategory",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CategoryCreateInput"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategorySuccessResponse"}}}
          },
          "401": {"description": "Unauthorized - Missing or invalid token"},
          "403": {"description": "Forbidden - Admin only"},
          "409": {"description": "Conflict - Duplicate category"},
          "500": {"description": "Internal server error"}
        }
      }
    },
    "/api/categories/hierarchy": {
      "get": {
        "tags": ["categories"],
        "summary": "Get all categories in hierarchical tree structure",
        "operationId": "getAllCategoriesHierarchy",
        "responses": {
          "200": {
            "description": "Categories hierarchy retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryHierarchyResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{id}": {
      "put": {
        "tags": ["categories"],
        "summary": "Update a category",
        "operationId": "updateCategory",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of category",
            "schema": {
              "type": "string",
              "example": "661fb4e9680c5e2a86f9f7b1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CategoryCreateInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategorySuccessResponse"}}}
          },
          "401": {"description": "Unauthorized - Missing or invalid token"},
          "403": {"description": "Forbidden - Admin only"},
          "404": {"description": "Category not found"},
          "500": {"description": "Internal server error"}
        }
      },
      "delete": {
        "tags": ["categories"],
        "summary": "Delete a category",
        "operationId": "deleteCategory",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of category",
            "schema": {
              "type": "string",
              "example": "661fb4e9680c5e2a86f9f7b1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategoryDeleteResponse"}}}
          },
          "401": {"description": "Unauthorized - Missing or invalid token"},
          "403": {"description": "Forbidden - Admin only"},
          "404": {"description": "Category not found"},
          "409": {"description": "Conflict - Cannot delete category with child categories"},
          "500": {"description": "Internal server error"}
        }
      }
    },
    "/api/categories/{id}/children": {
      "get": {
        "tags": ["categories"],
        "summary": "Get child categories",
        "operationId": "getChildCategories",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of parent category",
            "schema": {
              "type": "string",
              "example": "661fb4e9680c5e2a86f9f7b1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Child categories retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "example": "661fb4e9680c5e2a86f9f7b3"},
          "name": {"type": "string", "example": "Ăn vặt"},
          "parentId": {"type": ["string", "null"], "nullable": true, "example": null},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"}
        },
        "required": ["id", "name", "createdAt", "updatedAt"]
      },
      "CategoryCreateInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {"type": "string", "example": "Ăn vặt"},
          "parentId": {"type": "string", "nullable": true, "example": null}
        }
      },
      "CategorySuccessResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": true},
          "message": {"type": "string", "example": "Category retrieved successfully"},
          "data": {"$ref": "#/components/schemas/Category"},
          "timestamp": {"type": "string", "format": "date-time"}
        }
      },
      "CategoryListSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": true},
          "message": {"type": "string", "example": "Categories retrieved successfully"},
          "data": {"type": "array", "items": {"$ref": "#/components/schemas/Category"}},
          "timestamp": {"type": "string", "format": "date-time"}
        }
      },
      "CategoryDeleteResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": true},
          "message": {"type": "string", "example": "Category deleted successfully"},
          "data": {"type": "object", "properties": {"id": {"type": "string", "example": "661fb4e9680c5e2a86f9f7b1"}}},
          "timestamp": {"type": "string", "format": "date-time"}
        }
      },
      "CategoryNode": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "example": "661fb4e9680c5e2a86f9f7b1"},
          "name": {"type": "string", "example": "Cơm"},
          "child": {"type": "array", "items": {"$ref": "#/components/schemas/CategoryNode"}, "example": []}
        }
      },
      "CategoryHierarchyResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": true},
          "message": {"type": "string", "example": "Categories hierarchy retrieved successfully"},
          "data": {"type": "array", "items": {"$ref": "#/components/schemas/CategoryNode"}},
          "timestamp": {"type": "string", "format": "date-time"}
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "example": false},
          "message": {"type": "string", "example": "Failed to fetch users"},
          "error": {"nullable": true, "example": "Cast to ObjectId failed for value \"abc\""},
          "timestamp": {"type": "string", "format": "date-time"}
        },
        "required": ["success", "message", "timestamp"]
      },
...
    }
  }
}